stagelite/requirements.txt
amqp==2.5.2
appnope==0.1.0
asgiref==3.2.5
Babel==2.8.0
backcall==0.1.0
billiard==3.6.3.0
celery==4.4.2
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.2
Django==3.0.4
django-extensions==2.2.9
djangorestframework==3.11.0
flower==0.9.3
idna==2.9
importlib-metadata==1.5.2
ipython==7.14.0
ipython-genutils==0.2.0
jedi==0.17.0
kombu==4.6.8
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
praw==6.5.1
prawcore==1.0.1
prompt-toolkit==3.0.5
# psycopg2==2.8.4
ptyprocess==0.6.0
Pygments==2.6.1
pytz==2019.3
redis==3.4.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
tornado==5.1.1
traitlets==4.3.3
update-checker==0.16
urllib3==1.25.8
vine==1.3.0
wcwidth==0.1.9
websocket-client==0.57.0
zipp==3.1.0
